aws/lambda/aws-serverless-kafka-lambda-code/requirements.txt
kafka-python==2.2.7
aws-msk-iam-sasl-signer-python==1.0.2